Split examples into separate files and show source code
