+++ monadic-examples/src/main/scala/mhtml/examples/HelloWorld.scala
@@ -1,0 +1,5 @@
+package mhtml.examples
+
+object HelloWorld extends Example {
+  val app = <h2>Hello {"World"}!</h2>
+}
